CRUD+SWEETALER+VALIDATOR

+++ app/Http/Controllers/ProductController.php
@@ -9,6 +9,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Crypt;
 use Illuminate\Contracts\Encryption\DecryptException;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class ProductController extends Controller
 {
@@ -68,24 +69,47 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $values = request()->except('_token');
-        $productData = request()->except('_token');
-        $producto = new Product;
-        $producto->stock = $values['stock'];
-        $producto->name_product = $productData['name_product'];
-        $producto->description = $productData['description'];
-        $producto->price = $productData['price'];
-        if ($request->hasFile('image_product')) {
-            $producto->image_product = $request->file('image_product')->store('uploads', 'public');
-        }
-        $producto->id_category_product = $productData['id_category_product'];
-        $producto->save();
+        $rules = [
+            'name_product'          => 'required|string',
+            'stock'          => 'required|string',
+            'description'          => 'required|string',
+            'price'          => 'required|string',
+            'id_category_product'          => 'required|string',
+            'image_product' => 'required|file'
 
-        if ($request->hasFile('image_product')) {
-            $productData['image_product'] = $request->file('image_product')->store('uploads', 'public');
+        ];
+
+        $messages = [
+            'required'      => 'Este campo es obligatorio',
+        ];
+        $validator = Validator::make($request->all(), $rules, $messages);
+        if ($validator->passes()) {
+            DB::beginTransaction();
+            try {
+                $values = request()->except('_token');
+                $productData = request()->except('_token');
+                $producto = new Product;
+                $producto->stock = $values['stock'];
+                $producto->name_product = $productData['name_product'];
+                $producto->description = $productData['description'];
+                $producto->price = $productData['price'];
+                if ($request->hasFile('image_product')) {
+                    $producto->image_product = $request->file('image_product')->store('uploads', 'public');
+                }
+                $producto->id_category_product = $productData['id_category_product'];
+                $producto->save();
+                DB::connection(session()->get('database'))->commit();
+                return response('Se ingresó el producto con exito.', 200);
+            } catch (\Throwable $th) {
+                DB::connection(session()->get('database'))->rollBack();
+                return response('No se pudo realizar el ingreso del producto.', 400);
+            }
         }
 
-        return response('Se agregó el producto con exito.', 200);
+
+
+
+        return response('No se pudo realizar el ingreso del producto.', 400);
     }
 
     /**
@@ -158,24 +182,7 @@ class ProductController extends Controller
 
         // return view('mantenedores.product.edit', compact('productSelected'));
     }
-    public function productModalEditStore(request $request, product $product)
-    {
-        // $datosProducto = request()->except(['_token', '_method']);
-        // $producto = product::find($request->id);
-        // $producto->stock         = $request->stock;
-        // $producto->name_product   = $request->name_product;
-        // $producto->description    = $request->description;
-        // if ($request->image_product != null) {
-        //     Storage::delete('public/' . $request->urlImagen);
-        //     $producto->image_product = $request->file('image_product')->store('uploads', 'public');
-        // }
-        // $producto->save();
-        // return redirect()->route('product.index');
-        return json_encode(['HOL']);
 
-
-        // return view('mantenedores.product.edit', compact('productSelected'));
-    }
 
     /**
      * Update the specified resource in storage.
@@ -186,21 +193,47 @@ class ProductController extends Controller
      */
     public function update(request $request, product $product)
     {
-        $values = request()->except('_token');
-        dd($values);
-        $datosProducto = request()->except(['_token', '_method']);
-        $producto = product::find($product->id);
-        $producto->stock         = $request->stock;
-        $producto->name_product   = $request->name_product;
-        $producto->description    = $request->description;
-        $producto->id_category_product = $request->id_category_product;
-        $producto->price = $request->price;
-        if ($request->hasFile('image_product')) {
-            Storage::delete('public/' . $product->image_product);
-            $producto->image_product = $request->file('image_product')->store('uploads', 'public');
+
+        $rules = [
+            'name_product'          => 'required|string',
+            'stock'          => 'required|string',
+            'description'          => 'required|string',
+            'price'          => 'required|string',
+            'id_category_product'          => 'required|string',
+
+        ];
+
+        $messages = [
+            'required'      => 'Este campo es obligatorio',
+        ];
+
+
+        $validator = Validator::make($request->all(), $rules, $messages);
+        if ($validator->passes()) {
+            DB::beginTransaction();
+            try {
+
+                $producto = product::find($product->id);
+                $producto->stock         = $request->stock;
+                $producto->name_product   = $request->name_product;
+                $producto->description    = $request->description;
+                $producto->id_category_product = $request->id_category_product;
+                $producto->price = $request->price;
+                if ($request->hasFile('image_product')) {
+                    Storage::delete('public/' . $product->image_product);
+                    $producto->image_product = $request->file('image_product')->store('uploads', 'public');
+                }
+                $producto->save();
+                DB::connection(session()->get('database'))->commit();
+                return response('Se editó el producto con exito.', 200);
+            } catch (\Throwable $th) {
+                DB::connection(session()->get('database'))->rollBack();
+                return response('No se pudo editar el producto.', 400);
+            }
         }
-        $producto->save();
-        return redirect()->route('product.index');
+
+
+        return response('No se pudo editar el producto.', 400);
     }
 
     /**
@@ -211,17 +244,20 @@ class ProductController extends Controller
      */
     public function destroy(product $product)
     {
+
+        $id = $product->id;
         try {
-            $id = $product->id;
-        } catch (DecryptException $error) {
-            echo '<script>alert("Producto no eliminado")</script>';
-            return back();
+            $product = product::on(session()->get('database'))->find($id);
+            $product->delete();
+
+            DB::connection(session()->get('database'))->commit();
+        } catch (\Illuminate\Database\QueryException $e) {
+
+            DB::connection(session()->get('database'))->rollBack();
+
+            return response('Ocurrió un error. No se eliminó el producto.', 400);
         }
 
-        $product = product::on(session()->get('database'))->find($id);
-        $product->delete();
-
-
-        return redirect()->route('product.index');
+        return response('success', 200);
     }
 }
